libquic: do getsockopt(QUIC_SOCKOPT_TRANSPORT_PARAM_EXT) first
The content of it will not change, even after a
setsockopt(QUIC_SOCKOPT_TRANSPORT_PARAM_EXT),
because get/set operate on local/remote parameters.

So doing this first makes sure the socket is
actually a quic socket.

But the important thing is that it will
make refactoring easier.

The goal is to split quic_handshake() into
functions that don't operate on sockfd.

quic_handshake() itself will look like this:
{
    ret = quic_handshake_init(session);
    if (ret != 0) {
        return ret;
    }

    step = quic_handshake_next_step(session);
    while (step) {

        switch (step->op) {
        case GETSOCKOPT:
            // do getsockopt()
            break;
        case SETSOCKOPT:
            // do setsockopt()
            break;
        case SENDMSG:
            // do poll/sendmsg loop
            break;
        case RECVMSG:
            // do poll/recvmsg loop
            break;
        }

        /*
         * Do step operation on sockfd
         * getsockopt, setsockopt, sendmsg, recvmsg
         */
        ret = quic_handshake_process_step(session, step);
        if (ret != 0) {
            goto out;
        }

        step = quic_handshake_next_step(session);
    }

out:
    quic_handshake_deinit(session);
    return ret;
}

This will allow an application to do
a full async handshake and take care of
the interaction with sockfd itself.
E.g. this can interact with any
event loop the application is using.
Instead of having the poll based loops.
Even io_uring based versions of the
syscalls are possible.

gnutls does not handle GNUTLS_E_AGAIN
from the quic_tp_send() function.
So making the getsockopt() async from
within gnutls_handshake() is not possible.
So we do it before.

setsockopt() can be queued from within
quic_tp_recv(), similar to the queueing
of sendmsg() from quic_msg_read().

Signed-off-by: Stefan Metzmacher <metze@samba.org>